добавляет Route и заготовки страниц
