+++ src/pages/About/About.module.scss
@@ -1,0 +1,22 @@
+.root {
+    padding: 40px 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    h1, h2, p {
+        margin-bottom: 30px;
+    }
+
+
+    img {
+        display: block;
+        margin: 0 auto;
+        margin-bottom: 40px;
+    }  
+}
+
+.btnBack {
+    margin: 0 auto;
+    margin-bottom: 40px;
+}
