+++ src/pages/About/index.js
@@ -1,0 +1,3 @@
+import About from './About';
+
+export default About;
