+++ src/pages/Contacts/index.js
@@ -1,0 +1,3 @@
+import Contacts from './Contacts';
+
+export default Contacts;
