+++ src/pages/Main/index.js
@@ -1,0 +1,3 @@
+import Main from './Main';
+
+export default Main;
